Move specs to shared examples
