+++ spec/features/evidence_spec.rb
@@ -101,6 +101,10 @@ describe 'evidence' do
 
     it_behaves_like 'a form with a help button'
 
+    let(:autosaveable) { @evidence }
+    let(:path_params) { [current_project, @node, @evidence] }
+    it_behaves_like 'an editor with server side auto-save'
+
     describe 'textile form view' do
       let(:action_path) { edit_project_node_evidence_path(current_project, @node, @evidence) }
       let(:item) { @evidence }
