+++ spec/features/note_pages_spec.rb
@@ -99,6 +99,10 @@ describe "note pages" do
 
     it_behaves_like "a form with a help button"
 
+    let(:autosaveable) { @note }
+    let(:path_params) { [current_project, @node, @note] }
+    it_behaves_like 'an editor with server side auto-save'
+
     describe 'textile form view' do
       let(:action_path) { edit_project_node_note_path(current_project, @node, @note) }
       let(:item) { @note }
